move the files of back-end to a folder
